+++ packages/rainbowkit/src/wallets/connectorsForWallets.test.ts
@@ -2,7 +2,7 @@ import { describe, expect, it } from 'vitest';
 import { Connector } from 'wagmi';
 import { mainnet } from 'wagmi/chains';
 import { InjectedConnector } from 'wagmi/connectors/injected';
-import { WalletConnectConnector } from 'wagmi/connectors/walletConnect';
+import { WalletConnectLegacyConnector } from 'wagmi/connectors/walletConnectLegacy';
 import { connectorsForWallets } from '..';
 import { WalletInstance } from './Wallet';
 import { injectedWallet } from './walletConnectors';
@@ -27,7 +27,10 @@ describe('connectorsForWallets', () => {
           wallets: [
             {
               createConnector: () => ({
-                connector: new WalletConnectConnector({ chains, options: {} }),
+                connector: new WalletConnectLegacyConnector({
+                  chains,
+                  options: {},
+                }),
               }),
               iconBackground: '#fff',
               iconUrl: '/test.png',
@@ -49,7 +52,7 @@ describe('connectorsForWallets', () => {
 
       expect(getWalletInstances(connectors[1]).length).toBe(1);
       expect(getWalletInstances(connectors[1])[0].index).toBe(1);
-      expect(getWalletInstances(connectors[1])[0].name).toBe('Injected Wallet');
+      expect(getWalletInstances(connectors[1])[0].name).toBe('Browser Wallet');
     });
 
     it('includes injected fallback if no wallets using InjectedConnector are installed', () => {
@@ -70,7 +73,10 @@ describe('connectorsForWallets', () => {
             },
             {
               createConnector: () => ({
-                connector: new WalletConnectConnector({ chains, options: {} }),
+                connector: new WalletConnectLegacyConnector({
+                  chains,
+                  options: {},
+                }),
               }),
               iconBackground: '#fff',
               iconUrl: '/test.png',
@@ -98,7 +104,7 @@ describe('connectorsForWallets', () => {
 
       expect(getWalletInstances(connectors[2]).length).toBe(1);
       expect(getWalletInstances(connectors[2])[0].index).toBe(2);
-      expect(getWalletInstances(connectors[2])[0].name).toBe('Injected Wallet');
+      expect(getWalletInstances(connectors[2])[0].name).toBe('Browser Wallet');
     });
 
     it('includes injected fallback in the same order it was defined if no wallets using InjectedConnector are installed', () => {
@@ -125,7 +131,10 @@ describe('connectorsForWallets', () => {
           wallets: [
             {
               createConnector: () => ({
-                connector: new WalletConnectConnector({ chains, options: {} }),
+                connector: new WalletConnectLegacyConnector({
+                  chains,
+                  options: {},
+                }),
               }),
               iconBackground: '#fff',
               iconUrl: '/test.png',
@@ -158,7 +167,7 @@ describe('connectorsForWallets', () => {
 
       expect(getWalletInstances(connectors[2]).length).toBe(1);
       expect(getWalletInstances(connectors[2])[0].index).toBe(0); // Note this is now first in the list
-      expect(getWalletInstances(connectors[2])[0].name).toBe('Injected Wallet');
+      expect(getWalletInstances(connectors[2])[0].name).toBe('Browser Wallet');
       expect(getWalletInstances(connectors[2])[0].groupName).toBe(
         'Test Group 1'
       );
@@ -172,7 +181,9 @@ describe('connectorsForWallets', () => {
           wallets: [
             {
               createConnector: () => ({
-                connector: new InjectedConnector(),
+                connector: new InjectedConnector({
+                  options: { chains, name: 'Test Wallet Installed' },
+                }),
               }),
               iconBackground: '#fff',
               iconUrl: '/test.png',
@@ -188,7 +199,10 @@ describe('connectorsForWallets', () => {
             injectedWallet({ chains }),
             {
               createConnector: () => ({
-                connector: new WalletConnectConnector({ chains, options: {} }),
+                connector: new WalletConnectLegacyConnector({
+                  chains,
+                  options: {},
+                }),
               }),
               iconBackground: '#fff',
               iconUrl: '/test.png',