+++ packages/rainbowkit/src/wallets/getDefaultWallets.ts
@@ -6,6 +6,7 @@ import { coinbaseWallet } from './walletConnectors/coinbaseWallet/coinbaseWallet
 import { injectedWallet } from './walletConnectors/injectedWallet/injectedWallet';
 import { metaMaskWallet } from './walletConnectors/metaMaskWallet/metaMaskWallet';
 import { rainbowWallet } from './walletConnectors/rainbowWallet/rainbowWallet';
+import { safeWallet } from './walletConnectors/safeWallet/safeWallet';
 import { walletConnectWallet } from './walletConnectors/walletConnectWallet/walletConnectWallet';
 
 export const getDefaultWallets = ({
@@ -23,6 +24,7 @@ export const getDefaultWallets = ({
       groupName: 'Popular',
       wallets: [
         injectedWallet({ chains }),
+        safeWallet({ chains }),
         rainbowWallet({ chains }),
         coinbaseWallet({ appName, chains }),
         metaMaskWallet({ chains }),