+++ examples/with-remix/CHANGELOG.md
@@ -1,20 +1,87 @@
 # with-remix
 
-## null
+## 0.0.41
+
+### Patch Changes
+
+- Updated dependencies [9b93f56]
+  - @rainbow-me/rainbowkit@0.12.4
+
+## 0.0.40
+
+### Patch Changes
+
+- Updated dependencies [e7f1bec]
+- Updated dependencies [fe4f356]
+  - @rainbow-me/rainbowkit@0.12.3
+
+## 0.0.39
+
+### Patch Changes
+
+- Updated dependencies [2a1d230]
+- Updated dependencies [429a3c7]
+  - @rainbow-me/rainbowkit@0.12.2
+
+## 0.0.38
+
+### Patch Changes
+
+- Updated dependencies [3399df5]
+- Updated dependencies [8f01a12]
+  - @rainbow-me/rainbowkit@0.12.1
+
+## 0.0.37
+
+### Patch Changes
+
+- Updated dependencies [9838acf]
+  - @rainbow-me/rainbowkit@0.12.0
+
+## 0.0.36
+
+### Patch Changes
+
+- Updated dependencies [924ae82]
+- Updated dependencies [5e233ea]
+  - @rainbow-me/rainbowkit@0.11.1
+
+## 0.0.35
+
+### Patch Changes
+
+- Updated dependencies [1876ba0]
+  - @rainbow-me/rainbowkit@0.11.0
+
+## 0.0.34
+
+### Patch Changes
+
+- Updated dependencies [355402b]
+  - @rainbow-me/rainbowkit@0.10.0
+
+## 0.0.33
+
+### Patch Changes
+
+- Updated dependencies [49f0ec9]
+  - @rainbow-me/rainbowkit@0.9.0
+
+## 0.0.32
 
 ### Patch Changes
 
 - Updated dependencies [a1d6776]
   - @rainbow-me/rainbowkit@0.8.1
 
-## null
+## 0.0.31
 
 ### Patch Changes
 
 - Updated dependencies [6b37050]
   - @rainbow-me/rainbowkit@0.8.0
 
-## null
+## 0.0.30
 
 ### Patch Changes
 
@@ -22,21 +89,21 @@
 - Updated dependencies [0ff4210]
   - @rainbow-me/rainbowkit@0.7.4
 
-## null
+## 0.0.29
 
 ### Patch Changes
 
 - Updated dependencies [5a65178]
   - @rainbow-me/rainbowkit@0.7.3
 
-## null
+## 0.0.28
 
 ### Patch Changes
 
 - Updated dependencies [1de8203]
   - @rainbow-me/rainbowkit@0.7.2
 
-## null
+## 0.0.27
 
 ### Patch Changes
 
@@ -44,7 +111,7 @@
 - Updated dependencies [5ddc813]
   - @rainbow-me/rainbowkit@0.7.1
 
-## null
+## 0.0.26
 
 ### Patch Changes
 
@@ -52,7 +119,7 @@
 - Updated dependencies [2e6bb8f]
   - @rainbow-me/rainbowkit@0.7.0
 
-## null
+## 0.0.25
 
 ### Patch Changes
 
@@ -60,7 +127,7 @@
 - Updated dependencies [248a1cb]
   - @rainbow-me/rainbowkit@0.6.2
 
-## null
+## 0.0.24
 
 ### Patch Changes
 
@@ -68,7 +135,7 @@
 - Updated dependencies [fbf9d82]
   - @rainbow-me/rainbowkit@0.6.1
 
-## null
+## 0.0.23
 
 ### Patch Changes
 
@@ -79,14 +146,14 @@
 - Updated dependencies [c944ddc]
   - @rainbow-me/rainbowkit@0.6.0
 
-## null
+## 0.0.22
 
 ### Patch Changes
 
 - Updated dependencies [52e2ad6]
   - @rainbow-me/rainbowkit@0.5.3
 
-## null
+## 0.0.21
 
 ### Patch Changes
 
@@ -95,7 +162,7 @@
 - Updated dependencies [3f9013f]
   - @rainbow-me/rainbowkit@0.5.2
 
-## null
+## 0.0.20
 
 ### Patch Changes
 
@@ -104,7 +171,7 @@
 - Updated dependencies [8060ccd]
   - @rainbow-me/rainbowkit@0.5.1
 
-## null
+## 0.0.19
 
 ### Patch Changes
 
@@ -112,21 +179,21 @@
 - Updated dependencies [488c5a1]
   - @rainbow-me/rainbowkit@0.5.0
 
-## null
+## 0.0.18
 
 ### Patch Changes
 
 - Updated dependencies [4333995]
   - @rainbow-me/rainbowkit@0.4.8
 
-## null
+## 0.0.17
 
 ### Patch Changes
 
 - Updated dependencies [1a4f2f7]
   - @rainbow-me/rainbowkit@0.4.7
 
-## null
+## 0.0.16
 
 ### Patch Changes
 
@@ -134,21 +201,21 @@
 - Updated dependencies [948c036]
   - @rainbow-me/rainbowkit@0.4.6
 
-## null
+## 0.0.15
 
 ### Patch Changes
 
 - Updated dependencies [8dd5a74]
   - @rainbow-me/rainbowkit@0.4.5
 
-## null
+## 0.0.14
 
 ### Patch Changes
 
 - Updated dependencies [fd08aa1]
   - @rainbow-me/rainbowkit@0.4.4
 
-## null
+## 0.0.13
 
 ### Patch Changes
 
@@ -157,14 +224,14 @@
 - Updated dependencies [396308f]
   - @rainbow-me/rainbowkit@0.4.3
 
-## null
+## 0.0.12
 
 ### Patch Changes
 
 - Updated dependencies [0213b52]
   - @rainbow-me/rainbowkit@0.4.2
 
-## null
+## 0.0.11
 
 ### Patch Changes
 
@@ -172,28 +239,28 @@
 - Updated dependencies [3637bbb]
   - @rainbow-me/rainbowkit@0.4.1
 
-## null
+## 0.0.10
 
 ### Patch Changes
 
 - Updated dependencies [08d189b]
   - @rainbow-me/rainbowkit@0.4.0
 
-## null
+## 0.0.9
 
 ### Patch Changes
 
 - Updated dependencies [b2b46ef]
   - @rainbow-me/rainbowkit@0.3.7
 
-## null
+## 0.0.8
 
 ### Patch Changes
 
 - Updated dependencies [d905271]
   - @rainbow-me/rainbowkit@0.3.6
 
-## null
+## 0.0.7
 
 ### Patch Changes
 
@@ -202,28 +269,28 @@
 - Updated dependencies [1a7d50c]
   - @rainbow-me/rainbowkit@0.3.5
 
-## null
+## 0.0.6
 
 ### Patch Changes
 
 - Updated dependencies [ac63f9a]
   - @rainbow-me/rainbowkit@0.3.4
 
-## null
+## 0.0.5
 
 ### Patch Changes
 
 - Updated dependencies [ee81177]
   - @rainbow-me/rainbowkit@0.3.3
 
-## null
+## 0.0.4
 
 ### Patch Changes
 
 - Updated dependencies [33a2dd7]
   - @rainbow-me/rainbowkit@0.3.2
 
-## null
+## 0.0.3
 
 ### Patch Changes
 
@@ -231,14 +298,14 @@
 - Updated dependencies [11ed088]
   - @rainbow-me/rainbowkit@0.3.1
 
-## null
+## 0.0.2
 
 ### Patch Changes
 
 - Updated dependencies [233a6d7]
   - @rainbow-me/rainbowkit@0.3.0
 
-## null
+## 0.0.1
 
 ### Patch Changes